New DMROption 'layers' to produce layerwise DMR plots. (Changes made by Jaakko Moisio)
